MYR-16 Extract wiring helpers from main.go to fit CLAUDE.md 300-line cap
Per Claude Review suggestion #1: cmd/telemetry-server/main.go was 411
lines (pre-existing 388 worsened by this PR's +23). Extracted to a new
cmd/telemetry-server/wiring.go:

- setupEncryption (key load + Encryptor construction + startup log)
- setupAuthenticator (NoopAuth in dev / JWTAuth in prod)
- setupHTTPHandlers (Tesla, client, vehicle-status, fleet-config, debug-fields)
- setupTeslaTLS (cert/key load + TLS config)
- setupFleetConfigEndpoint (moved from main.go)
- buildTeslaTLS (moved from main.go)

main.go drops from 411 → 248 lines. wiring.go is 241 lines. Both well
under the cap.

Pure refactor — no behavior change. Cyclop suppression dropped from run()
since it's now within the threshold; funlen kept because the file is
still close to the line cap and the composition root naturally needs the
sequential wiring layout.
